refactor(soundpacks): split wizard shell state
Move sound pack wizard state, shell widgets, and scrolled-panel compatibility into focused helpers while keeping step behavior in the dialog.

Tested: pytest tests/test_soundpack_paths.py; ruff check src tests scripts; pyright.

Confidence: high
